# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
...
  /coupons/validate:
    post:
      operationId: validateCouponCode
      tags: [coupons]
      summary: Validate a coupon code and preview its discount for a given product/duration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponValidateInput"
      responses:
        "200":
          description: Validation result (ok=false explains why the coupon can't be used)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidateResponse"
...
  /admin/coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all coupons (admin)
      responses:
        "200":
          description: List of coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCoupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create a coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCoupon"
        "400":
          description: Validation error
  /admin/coupons/{id}:
    put:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update a coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CouponInput"
      responses:
        "200":
          description: Coupon updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCoupon"
        "400":
          description: Validation error
        "404":
          description: Not found
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete a coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /admin/coupons/{id}/redemptions:
    get:
      operationId: listCouponRedemptions
      tags: [coupons]
      summary: List redemptions for a coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of redemptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CouponRedemption"
...
components:
  schemas:
...
    OrderInput:
      type: object
      required: [productId, customerEmail, customerName]
      properties:
        productId:
          type: integer
        customerEmail:
          type: string
        customerName:
          type: string
        durationMonths:
          type: integer
          description: "1, 3, or 12 — defaults to 1"
        couponCode:
          type: ["string", "null"]
          description: Optional coupon code to apply to this order.
        useStoreCredit:
          type: boolean
          description: When true, applies as much of the buyer's available store credit as possible.
    Order:
      type: object
      required: [id, productId, customerEmail, customerName, amountKobo, status, reference, createdAt, durationMonths]
      properties:
        id:
          type: integer
        productId:
          type: integer
        customerEmail:
          type: string
        customerName:
          type: string
        amountKobo:
          type: integer
        status:
          type: string
        reference:
          type: string
        createdAt:
          type: string
        durationMonths:
          type: integer
        discountKobo:
          type: integer
          description: Amount discounted by a coupon, if any.
        couponCode:
          type: ["string", "null"]
        creditAppliedKobo:
          type: integer
          description: Store credit applied to this order, if any.
...
    CouponValidateInput:
      type: object
      required: [code, productId]
      properties:
        code:
          type: string
        productId:
          type: integer
        durationMonths:
          type: integer
          description: "1, 3, or 12 — defaults to 1"
    CouponValidateResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        error:
          type: string
        discountKobo:
          type: integer
        code:
          type: string
    AdminCoupon:
      type: object
      required:
        [id, code, discountType, discountValue, scope, productIds, minPurchaseKobo, usageLimitTotal,
         usageLimitPerCustomer, usedCount, requiresLogin, active, createdAt, updatedAt]
      properties:
        id:
          type: integer
        code:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
        discountValue:
          type: integer
        scope:
          type: string
        productIds:
          type: array
          items:
            type: integer
        minPurchaseKobo:
          type: integer
        maxDiscountKobo:
          type: ["integer", "null"]
        usageLimitTotal:
          type: ["integer", "null"]
        usageLimitPerCustomer:
          type: ["integer", "null"]
        usedCount:
          type: integer
        requiresLogin:
          type: boolean
        active:
          type: boolean
        startsAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        createdBy:
          type: ["string", "null"]
    CouponInput:
      type: object
      required: [code, discountType, discountValue]
      properties:
        code:
          type: string
        description:
          type: ["string", "null"]
        discountType:
          type: string
          description: "percentage | fixed"
        discountValue:
          type: integer
        scope:
          type: string
          description: "all | selected"
        productIds:
          type: array
          items:
            type: integer
        minPurchaseKobo:
          type: integer
        maxDiscountKobo:
          type: ["integer", "null"]
        usageLimitTotal:
          type: ["integer", "null"]
        usageLimitPerCustomer:
          type: ["integer", "null"]
        requiresLogin:
          type: boolean
        active:
          type: boolean
        startsAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]
    CouponRedemption:
      type: object
      required: [id, couponId, customerEmail, discountKobo, createdAt]
      properties:
        id:
          type: integer
        couponId:
          type: integer
        orderId:
          type: ["integer", "null"]
        customerEmail:
          type: string
        clerkUserId:
          type: ["string", "null"]
        discountKobo:
          type: integer
        createdAt:
          type: string
...