# lib/api-zod/src/generated/types/adminCoupon.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCoupon {
  id: number;
  code: string;
  /** @nullable */
  description?: string | null;
  discountType: string;
  discountValue: number;
  scope: string;
  productIds: number[];
  minPurchaseKobo: number;
  /** @nullable */
  maxDiscountKobo?: number | null;
  /** @nullable */
  usageLimitTotal: number | null;
  /** @nullable */
  usageLimitPerCustomer: number | null;
  usedCount: number;
  requiresLogin: boolean;
  active: boolean;
  /** @nullable */
  startsAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  createdBy?: string | null;
}