# lib/api-zod/src/generated/types/couponInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CouponInput {
  code: string;
  /** @nullable */
  description?: string | null;
  /** percentage | fixed */
  discountType: string;
  discountValue: number;
  /** all | selected */
  scope?: string;
  productIds?: number[];
  minPurchaseKobo?: number;
  /** @nullable */
  maxDiscountKobo?: number | null;
  /** @nullable */
  usageLimitTotal?: number | null;
  /** @nullable */
  usageLimitPerCustomer?: number | null;
  requiresLogin?: boolean;
  active?: boolean;
  /** @nullable */
  startsAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}