# lib/api-zod/src/generated/types/couponRedemption.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CouponRedemption {
  id: number;
  couponId: number;
  /** @nullable */
  orderId?: number | null;
  customerEmail: string;
  /** @nullable */
  clerkUserId?: string | null;
  discountKobo: number;
  createdAt: string;
}