# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Order {
  id: number;
  productId: number;
  customerEmail: string;
  customerName: string;
  amountKobo: number;
  status: string;
  reference: string;
  createdAt: string;
  durationMonths: number;
  /** Amount discounted by a coupon, if any. */
  discountKobo?: number;
  /** @nullable */
  couponCode?: string | null;
  /** Store credit applied to this order, if any. */
  creditAppliedKobo?: number;
}