# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  productId: number;
  customerEmail: string;
  customerName: string;
  /** 1, 3, or 12 — defaults to 1 */
  durationMonths?: number;
  /**
     * Optional coupon code to apply to this order.
     * @nullable
     */
  couponCode?: string | null;
  /** When true, applies as much of the buyer's available store credit as possible. */
  useStoreCredit?: boolean;
}